Rename go module  path
