+++ go/internal/store/projects-memory.go
@@ -3,8 +3,8 @@ package store
 import (
 	"crypto/rand"
 
-	"github.com/spfave/projects-build/apps/server-go/internal/core"
-	pErr "github.com/spfave/projects-build/apps/server-go/pkg/errors"
+	"github.com/spfave/projects-build/go/internal/core"
+	pErr "github.com/spfave/projects-build/go/pkg/errors"
 )
 
 type ProjectMemoryStore struct {
